update artifact directories

+++ bin/RQ3/sparrow.py
@@ -5,12 +5,10 @@ import csv
 import config
 import datetime
 import subprocess
-from logger import log, INFO, ERROR, WARNING
+from logger import log, INFO, ERROR, WARNING, ALL
 from typing import TextIO
 import copy
 import time
-SPARROW_LOG_DIR = os.path.join(config.configuration['OUT_DIR'], 'sparrow_logs')
-IS_SOLO_SPARROW = False
 
 '''
 Function that actually runs Sparrow
@@ -21,13 +19,31 @@ Output: tuple (TextIO, subprocess.Popen)
 '''
 def run_sparrow(file: str) -> tuple[TextIO, subprocess.Popen]:
     os.chdir(os.path.dirname(file))
-    sparrow_log = open('sparrow_log', 'w')
-    if os.path.exists(os.path.join(os.path.dirname(file), 'sparrow-out')):
-        os.system(f'rm -rf {os.path.join(os.path.dirname(file), "sparrow-out")}')
-    if IS_SOLO_SPARROW or config.configuration["USER_SPARROW_OPT"] == []:
-        cmd = [config.configuration["SPARROW_BIN_PATH"], file] + config.configuration["DEFAULT_SPARROW_OPT"] + config.configuration["ADDITIONAL_SPARROW_OPT"]
-    else:
-        cmd = [config.configuration["SPARROW_BIN_PATH"], file] + config.configuration["USER_SPARROW_OPT"] + config.configuration["ADDITIONAL_SPARROW_OPT"]
+    match config.configuration["PURPOSE"]:
+        case "PREPROCESS":
+            if os.path.exists(os.path.join(os.path.dirname(file), 'sparrow-out')) and not config.configuration["OVERWRITE_SPARROW"]:
+                log(ALL, f"{os.path.join(os.path.dirname(file), 'sparrow-out')} already exists.")
+                return None, None
+            else:
+                try:
+                    os.system(f'rm -rf {os.path.join(os.path.dirname(file), "sparrow-out")}')
+                except:
+                    log(ERROR, f"Failed to remove {os.path.join(os.path.dirname(file), 'sparrow-out')}")
+                    return None, None
+                sparrow_log = open('sparrow_log', 'w')
+            cmd = [config.configuration["SPARROW_BIN_PATH"], file] + config.preprocess_configuration["DEFAULT_SPARROW_OPT"] + config.preprocess_configuration["USER_SPARROW_OPT"]
+        case "SPARROW":
+            if os.path.exists(os.path.join(os.path.dirname(file), 'sparrow-out')) and not config.sparrow_configuration["OVERWRITE_SPARROW"]:
+                log(ALL, f"{os.path.join(os.path.dirname(file), 'sparrow-out')} already exists.")
+                return None, None
+            else:
+                try:
+                    os.system(f'rm -rf {os.path.join(os.path.dirname(file), "sparrow-out")}')
+                except:
+                    log(ERROR, f"Failed to remove {os.path.join(os.path.dirname(file), 'sparrow-out')}")
+                    return None, None
+                sparrow_log = open('sparrow_log', 'w')
+            cmd = [config.configuration["SPARROW_BIN_PATH"], file] + config.sparrow_configuration["DEFAULT_SPARROW_OPT"] + config.sparrow_configuration["USER_SPARROW_OPT"]
     log(INFO, f"Running sparrow with {cmd}")
     return sparrow_log, subprocess.Popen(cmd,
                                  stdout=sparrow_log,
@@ -42,12 +58,10 @@ Output: bool (True: success, False: fail)
 def sparrow(package:str, files:list) -> bool:
     start_time = time.time()
     time_record = dict()
-    SPARROW_PKG_DIR = os.path.join(SPARROW_LOG_DIR, package)
-    if not os.path.exists(SPARROW_LOG_DIR):
-        os.mkdir(SPARROW_LOG_DIR)
+    SPARROW_PKG_DIR = os.path.join(config.configuration["SPARROW_LOG_DIR"], package)
     if not os.path.exists(SPARROW_PKG_DIR):
         os.mkdir(SPARROW_PKG_DIR)
-    tsvfile = open(os.path.join(SPARROW_LOG_DIR, '{}_sparrow_stat.tsv'.format(package)), 'a')
+    tsvfile = open(os.path.join(config.configuration["SPARROW_LOG_DIR"], '{}_sparrow_stat.tsv'.format(package)), 'a')
     writer = csv.writer(tsvfile, delimiter='\t')
     writer.writerow(['Package', 'Status'])
     tsvfile.flush()
@@ -61,6 +75,11 @@ def sparrow(package:str, files:list) -> bool:
         time_record[files[i]] = dict()
         time_record[files[i]]['start'] = time.time()
         sparrow_log, process = run_sparrow(files[i])
+        if sparrow_log is None or process is None:
+            log(ALL, f"Skipping {files[i]} ...")
+            rest_files.remove(files[i])
+            i += 1
+            continue
         proc_cnt += 1
         procs.append((files[i], process, sparrow_log))
         rest_files.remove(files[i])
@@ -85,7 +104,7 @@ def sparrow(package:str, files:list) -> bool:
                     continue
                 sparrow_log.close()
                 if process.returncode == 0:
-                    log(INFO, f"{file} is successfully analyzed.")
+                    log(ALL, f"{file} is successfully analyzed.")
                     time_record[file]['end'] = time.time()
                     writer.writerow([file, 'O'])
                     success_cnt += 1
@@ -107,49 +126,74 @@ def sparrow(package:str, files:list) -> bool:
     time_lst = []
     with open(os.path.join(SPARROW_PKG_DIR, f'{package}_time_summary.txt'), 'w') as f:
         for file in time_record:
-            # check if 'end' key exists
             if 'end' not in time_record[file] or time_record[file]['end'] == 0:
-                f.write(f"{file}: Timeout or Error\n")
+                f.write(f"{file}: Finished too quickly or Timeout or Error\n")
             else:
                 time_lst.append(time_record[file]['end'] - time_record[file]['start'])
                 f.write(f"{file}: {time_record[file]['end'] - time_record[file]['start']}\n")
         if len(time_lst) != 0:
             f.write(f"Average: {sum(time_lst)/len(time_lst)}\n")
         f.write(f"Total time: {time.time() - start_time}\n")
-    log(INFO, f"{success_cnt} files are successfully analyzed.")
+    log(ALL, f"{success_cnt} files are successfully analyzed.")
+    log(ALL, f"Output for analysis is saved in {SPARROW_PKG_DIR}")
     return True
 
-'''
-Function called from -pipe option of bin/oss_exp.py
+# '''
+# Function called from -pipe option of bin/oss_exp.py
 
-Input: str (package name), TextIO (tsvfile), csv.writer
-Output: bool (True: success, False: fail)
-'''
-def sparrow_pipe(package:str, tsvfile:TextIO, writer:csv.writer) -> bool:
-    global SPARROW_LOG_DIR
-    SPARROW_LOG_DIR = os.path.join(config.configuration['OUT_DIR'], 'sparrow_logs')
-    if not os.path.exists(SPARROW_LOG_DIR):
-        os.mkdir(SPARROW_LOG_DIR)
-    target_dir = os.path.join(config.configuration["ANALYSIS_DIR"], package)
-    if not os.path.exists(target_dir):
-        log(ERROR, f"{target_dir} does not exist.")
-        writer.writerow([package, 'O', 'O', 'X', '-', "dir not found"])
-        tsvfile.flush()
-        return False
-    c_files = []
-    for root, dirs, files in os.walk(target_dir):
-        for file in files:
-            if file.endswith('.c'):
-                c_files.append(os.path.join(root, file))
-    if len(c_files) == 0:
-        log(ERROR, f"No .c files found in {target_dir}.")
-        writer.writerow([package, 'O', 'O', 'X', '-', "no .c file found"])
-        tsvfile.flush()
-        return False
-    log(INFO, f"Found {len(c_files)} .c files in {target_dir}.")
-    return sparrow(package, c_files)
+# Input: str (package name), TextIO (tsvfile), csv.writer
+# Output: bool (True: success, False: fail)
+# '''
+# def sparrow_pipe(package:str, tsvfile:TextIO, writer:csv.writer) -> bool:
+#     global SPARROW_LOG_DIR
+#     config.configuration["SPARROW_LOG_DIR"] = os.path.join(config.configuration['OUT_DIR'], 'sparrow_logs')
+#     if not os.path.exists(SPARROW_LOG_DIR):
+#         os.mkdir(SPARROW_LOG_DIR)
+#     target_dir = os.path.join(config.configuration["ANALYSIS_DIR"], package)
+#     if not os.path.exists(target_dir):
+#         log(ERROR, f"{target_dir} does not exist.")
+#         writer.writerow([package, 'O', 'O', 'X', '-', "dir not found"])
+#         tsvfile.flush()
+#         return False
+#     c_files = []
+#     for root, dirs, files in os.walk(target_dir):
+#         for file in files:
+#             if file.endswith('.c'):
+#                 c_files.append(os.path.join(root, file))
+#     if len(c_files) == 0:
+#         log(ERROR, f"No .c files found in {target_dir}.")
+#         writer.writerow([package, 'O', 'O', 'X', '-', "no .c file found"])
+#         tsvfile.flush()
+#         return False
+#     log(INFO, f"Found {len(c_files)} .c files in {target_dir}.")
+#     return sparrow(package, c_files)
                 
-                
+def mk_worklist(targets):
+    log(ALL, "Looking for .c files in the target directories ...")
+    target_files = []
+    for target in targets:
+        for root, dirs, files in os.walk(target):
+            for file in files:
+                if file.endswith('.c'):
+                    target_files.append(os.path.abspath(os.path.join(root, file)))
+    log(ALL, f"Found {len(target_files)} .c files for the analysis target.")
+    target_sorted = dict()
+    for target in target_files:
+        if_found_package_name = False
+        for tok in target.split('/'):
+            if tok.startswith('_'):
+                target_sorted[tok] = [target] if tok not in target_sorted else target_sorted[tok] + [target]
+                if_found_package_name = True
+                break
+        if not if_found_package_name:
+            target_sorted['UNKNOWN'] = [target] if 'UNKNOWN' not in target_sorted else target_sorted['UNKNOWN'] + [target]
+    return target_sorted
+            
+def execute_worklist(worklist):
+    log(ALL, f"@@@WARNING: ANALYSIS PROCESS MIGHT TAKE UNEXPECTEDLY LONG TIME DEPENDING ON THE # ALARMS@@@")
+    for key in worklist:
+        log(ALL, f"Analyzing on {key}, # of files: {len(worklist[key])}")
+        sparrow(key, worklist[key])
 '''
 Run this script directly with -f option (target directory(s))
 -f argument must have one or more .c files for Sparrow to analyze
@@ -169,20 +213,14 @@ Input: str (package name), TextIO (tsvfile), csv.writer
 Output: bool (True: success, False: fail)
 '''
 def main():
-    IS_SOLO_SPARROW = True
-    config.setup("SPARROW")
-    log(INFO, "You are running sparrow.py script directly.")
-    log(INFO, "This script only runs on the first argument of -f option.")
-    target_files = []
-    if not os.path.exists(config.configuration["ARGS"].files[0]):
-        log(ERROR, f"{config.configuration['ARGS'].files[0]} does not exist.")
-        config.patron_exit("SPARROW")
-    for root, dirs, files in os.walk(config.configuration["ARGS"].files[0]):
-        for file in files:
-            if file.endswith('.c'):
-                target_files.append(os.path.abspath(os.path.join(root, file)))
-    print(config.configuration["ARGS"].files[0])
-    sparrow("top", target_files)
+    for d in config.configuration["ARGS"].target_directory:
+        if not os.path.exists(d):
+            log(ERROR, f"Given directory: {d} does not exist.")
+            config.bad_ending(config.configuration["OUT_DIR"])
+    worklist = mk_worklist(config.configuration["ARGS"].target_directory)
+    execute_worklist(worklist)
+    config.happy_ending(config.configuration["OUT_DIR"])
 
 if __name__ == '__main__':
+    config.setup_sparrow()
     main()
